pyproject.toml
[project]
name = "tier3-gcp-cloud-run"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.1",
  "google-cloud-secret-manager>=2.24.0",
  "httpx>=0.28.1",
  "openai>=1.97.0",
  "pydantic>=2.11.7",
  "python-dotenv>=1.1.1",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.35.0",
]
